implemented CRs
