+++ cli/src/test/java/com/devonfw/tools/ide/context/AbstractIdeTestContext.java
@@ -29,6 +29,8 @@ public class AbstractIdeTestContext extends AbstractIdeContext {
 
   private FileAccess mockFileAccess;
 
+  private Path dummyUserHome;
+
   /**
    * The constructor.
    *
@@ -97,11 +99,33 @@ public class AbstractIdeTestContext extends AbstractIdeContext {
     this.systemInfo = systemInfo;
   }
 
-
   /**
    * @param fileAccess the {@link FileAccess} to use for testing.
    */
-  public void setMockFileAccess(FileAccess fileAccess){
+  public void setMockFileAccess(FileAccess fileAccess) {
+
     this.mockFileAccess = fileAccess;
+  }
+
+  /**
+   * @param dummyUserHome mock path which will be used in {@link #getUserHome()}
+   */
+  public void setUserHome(Path dummyUserHome) {
+
+    this.dummyUserHome = dummyUserHome;
+  }
+
+  /**
+   * @return a dummy UserHome path to avoid global path access in a commandlet test. The defined {@link #dummyUserHome} will be returned if it is not
+   * {@code null}, else see implementation {@link #AbstractIdeContext}.
+   */
+  @Override
+  public Path getUserHome() {
+
+    if (dummyUserHome != null) {
+      return dummyUserHome;
+    }
+
+    return super.getUserHome();
   }
 }
